Rename module to representor
