+++ tests/adapters/maze_xml_test.py
@@ -2,8 +2,8 @@ import xml.etree.ElementTree as ET
 import unittest
 import json
 
-from hypermedia_resource import HypermediaResource
-from hypermedia_resource.contrib.maze_xml import MazeXMLAdapter
+from representor import HypermediaResource
+from representor.contrib.maze_xml import MazeXMLAdapter
 
 cell_xml = """<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
 <maze version="1.0">
